notebooks/chroma-embeddings/requirements.txt
chromadb
openai
tiktoken
python-dotenv